Develop

+++ 3rdparty/boost/libs/safe_numerics/example/example17.cpp
@@ -1,7 +1,14 @@
+//  Copyright (c) 2018 Robert Ramey
+//
+// Distributed under the Boost Software License, Version 1.0. (See
+// accompanying file LICENSE_1_0.txt or copy at
+// http://www.boost.org/LICENSE_1_0.txt)
+
 #include <boost/safe_numerics/safe_integer.hpp>
+#include <cstdint> // uint8_t
 using namespace boost::safe_numerics;
 
-int f(int i){
+uint8_t f(uint8_t i){
     return i;
 }
 
@@ -9,8 +16,8 @@ using safe_t = safe<long, native, loose_trap_policy>;
 
 int main(){
     const long x = 97;
-    f(x);   // OK - implicit conversion to int
+    f(x);   // OK - implicit conversion to int can never fail
     const safe_t y = 97;
-    f(y);   // Would be OK, but will invoke compile time error
+    f(y);   // could overflow so trap at compile time
     return 0;
 }
